Time: 0 ms (100%), Space: 20 MB (66.85%) - LeetHub

+++ 0122-best-time-to-buy-and-sell-stock-ii/0122-best-time-to-buy-and-sell-stock-ii.cpp
@@ -1,13 +1,13 @@
 class Solution {
 public:
     int maxProfit(vector<int>& prices) {
-        vector<long> ahead(2,0);
-        vector<long> curr(2,0);
+        vector<int> ahead(2,0);
+        vector<int> curr(2,0);
         ahead[0] = ahead[1] = 0;
         int n = prices.size();
         for(int i = n-1;i>=0;i--){
             for(int buy =0;buy<=1;buy++){
-                long profit = 0;
+                int profit = 0;
                 if(buy){
                     profit = max(-prices[i]+ahead[0],ahead[1]);
                 } else {
